+++ src/main/java/com/example/webtemplate/account/AccountNotFoundAdvice.java
@@ -1,10 +1,12 @@
-package com.example.webtemplate;
+package com.example.webtemplate.account;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import com.example.webtemplate.Util;
 
 @RestControllerAdvice
 class AccountNotFoundAdvice {
