+++ src/main/java/com/example/webtemplate/account/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.example.webtemplate;
+package com.example.webtemplate.account;
 
 import org.springframework.data.jpa.repository.JpaRepository;
