+++ src/main/java/com/example/webtemplate/account/LoadDatabase.java
@@ -1,4 +1,4 @@
-package com.example.webtemplate;
+package com.example.webtemplate.account;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
